+++ src/libxtp/qmpackages/orca.cc
@@ -272,7 +272,7 @@ bool Orca::WriteInputFile(const Orbitals& orbitals) {
   for (const auto& prop : this->_settings.property("orca")) {
     const std::string& prop_name = prop.name();
     if (prop_name == "pointcharges") {
-      _options += this->CreateInputSection("orca.pointcharges", true);
+      _options += this->CreateInputSection("orca.pointcharges");
     } else if (prop_name != "method") {
       _options += this->CreateInputSection("orca." + prop_name);
     }
@@ -803,13 +803,12 @@ std::string Orca::indent(const double& number) {
   return snumber;
 }
 
-std::string Orca::CreateInputSection(const std::string& key,
-                                     bool single_line) const {
+std::string Orca::CreateInputSection(const std::string& key) const {
   std::stringstream stream;
   std::string section = key.substr(key.find(".") + 1);
   stream << "%" << section;
-  if (single_line) {
-    stream << " " << _settings.get(key) << "\n";
+  if (KeywordIsSingleLine(key)) {
+    stream << " " << this->_settings.get(key) << "\n";
   } else {
     stream << "\n"
            << this->_settings.get(key) << "\n"
@@ -817,6 +816,13 @@ std::string Orca::CreateInputSection(const std::string& key,
   }
 
   return stream.str();
+}
+
+bool Orca::KeywordIsSingleLine(const std::string& key) const {
+  tools::Tokenizer values(this->_settings.get(key), " ");
+  std::vector<std::string> words;
+  values.ToVector(words);
+  return ((words.size() <= 1) ? true : false);
 }
 
 std::string Orca::WriteMethod() const {